Simplify ivy-avy and swiper-avy key bindings
Both counsel-ag and counsel-grep seem to work with the default C-'
binding of ivy-avy in ivy-minibuffer-map, so overriding it with
swiper-avy may be unnecessary (#1711).

Meanwhile swiper works only with swiper-avy and not ivy-avy, so
remap the latter to the former in swiper-map.  This means users who
change the C-' binding in ivy-minibuffer-map will not have to also
repeat the same in swiper-map (#1711, #3074).  Keep C-' override for
backward compatibility.

* counsel.el (counsel-ag-map, counsel-grep-map): Remove swiper-avy
key binding.  Add docstring.
* swiper.el (swiper-map): Remap ivy-avy to swiper-avy.